+++ frontends/p4/specializeGenericTypes.cpp
@@ -15,6 +15,7 @@ limitations under the License.
 */
 
 #include "specializeGenericTypes.h"
+#include "frontends/p4/typeChecking/typeSubstitutionVisitor.h"
 
 namespace P4 {
